+++ trunk/src/gl/gl_skydome.cpp
@@ -188,10 +188,10 @@ static PalEntry SkyCapColor(FTextureID texno, bool bottom)
 
 				if (buffer)
 				{
-					SkyColors[texno.GetIndex()]=averageColor((unsigned long *) buffer, w * MIN(30, h), false);
+					SkyColors[texno.GetIndex()]=averageColor((DWORD *) buffer, w * MIN(30, h), false);
 					if (h>30)
 					{
-						SkyColors[texno.GetIndex()+MaxSkyTexture] = averageColor(((unsigned long *) buffer)+(h-30)*w, w * 30, false);
+						SkyColors[texno.GetIndex()+MaxSkyTexture] = averageColor(((DWORD *) buffer)+(h-30)*w, w * 30, false);
 					}
 					else SkyColors[texno.GetIndex() + MaxSkyTexture] = SkyColors[texno.GetIndex()];
 					delete buffer;
